Completed POC with Logger

+++ src/main/java/com/rohit/controller/LambdaHandler.java
@@ -4,13 +4,14 @@ import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
-import com.rohit.exceptions.userExceptions;
-import com.rohit.services.Services;
-//import  software.amazon.lambda.powertools.validation.Validation;
-//import software.amazon.lambda.powertools.validation.ValidationException;
 
+import com.rohit.services.Services;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class LambdaHandler implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
+    static final Logger logger = LogManager.getLogger(LambdaHandler.class);
+
 
     //@Validation(inboundSchema = "classpath:/schema.json")
     public APIGatewayProxyResponseEvent handleRequest(APIGatewayProxyRequestEvent request, Context context)
@@ -18,23 +19,42 @@ public class LambdaHandler implements RequestHandler<APIGatewayProxyRequestEvent
         Services service = new Services();
 
             //ValidationUtils.validate(request.getBody(),"classpath:/schema.json");
-            String method = request.getHttpMethod();
-            switch (method) {
-                case "POST":
-                    return service.saveEmployee(request, context);
-                case "GET":
-                    if (request.getPathParameters() != null) {
-                        return service.getEmployeeById(request, context);
+            if(service.checkForLogin(request,context))
+            {
+                String method = request.getHttpMethod();
+                    switch (method) {
+                        case "POST":
+                            return service.validateUser(request,context);
+                        default:
+                            logger.fatal("UnSupportedOperationException triggered for Login");
+                            return service.ErrorResponse("UnSupportedOperationException triggered for Login",405);
                     }
-                    return service.getEmployees(request, context);
 
-                case "DELETE":
-                    return service.deleteById(request, context);
-                default:
-                    throw new Error("Unsupported Methods:::" + request.getHttpMethod());
+            }else{
+                String method = request.getHttpMethod();
+                    switch (method) {
+                        case "POST":
+                            if (service.validateInput(request, context)) {
+                                return service.saveEmployee(request, context);
+                            } else {
+                                logger.fatal("InvalidPropertiesFormatException triggered");
+                                return service.ErrorResponse("InvalidPropertiesFormatException triggered for Register",406);
+
+                            }
+
+                        case "GET":
+                            if (request.getPathParameters() != null) {
+                                return service.getEmployeeById(request, context);
+                            }
+                            return service.getEmployees(request, context);
+
+                        case "DELETE":
+                            return service.deleteById(request, context);
+                        default:
+                            logger.fatal("UnSupportedOperationException triggered");
+                            return service.ErrorResponse("UnSupportedOperationException triggered",405);
+                    }
             }
+        }
     }
 
-
-
-}
